requirements.txt
numpy
pandas
scipy
matplotlib
Pillow
opencv-python
notebook
